cambios varios
-revision de entidades
-revision repositorios
-revision controllers
-revision persistencia
-instalacion swagger

+++ services-api/services-api-service/src/main/java/com/ceiba/services/api/service/UsuarioService.java
@@ -2,8 +2,8 @@ package com.ceiba.services.api.service;
 
 import org.springframework.stereotype.Service;
 
-import com.ceiba.services.api.common.dto.RequestService;
-import com.ceiba.services.api.common.dto.ResponseService;
+import com.ceiba.services.api.service.common.dto.RequestService;
+import com.ceiba.services.api.service.common.dto.ResponseService;
 import com.ceiba.services.api.service.port.in.UsuarioUseCase;
 import com.ceiba.services.api.service.port.out.UsuarioPort;
 
@@ -18,7 +18,7 @@ public class UsuarioService implements UsuarioUseCase {
 	public ResponseService consultar(RequestService req) {
 		ResponseService resp = new ResponseService();
 		if (!req.getId().isEmpty()) {
-			resp.setUsuario(usuarioPort.getUsuario(Long.parseLong(req.getId())));
+			resp.setUsuario(usuarioPort.getUsuario(Integer.parseInt(req.getId())));
 		} else if (!req.getNombre().isEmpty()) {
 			resp.setUsuarios(usuarioPort.getUsuariosByNombre(req.getNombre()));
 		} else
